Initial commit for NINA-B1 support on the iMX8M SOMS
This includes hardware specific configuration changes and
dependencies for building an flashing firmware on the
integrated NINA-B1 module on the iMX8M SOMs, and carrier
boards.  The default configuration is for the configuration
of the HummingBoard Pulse.

This BSP has been used to build the blehci app for integrating
the NINA-B1 into the default Linux Bluetooth stack.

Signed-off-by: Jon Nettleton <jon@solid-run.com>